cats_vs_dogs/backend/requirements.txt
fastapi
uvicorn
requests
pillow
numpy
python-multipart
